Implement remove-book feature
